move file operation utils to localstack.utils.files
